create NCPO when processingStatus is met
