+++ app/Http/Controllers/RequestNcpoController.php
@@ -25,40 +25,40 @@ class RequestNcpoController extends Controller
         ]);
     }
 
-    public function store(StoreNcpoRequest $request)
-    {
-        $validated = $request->validated();
-        $ncpo = DB::transaction(function () use ($validated) {
-            $ncpoNo = RequestNCPO::generateReferenceNumber(
-                'ncpo_no',
-                fn ($prefix, $datePart, $number) => "{$prefix}-{$datePart}-{$number}",
-                ['prefix' => 'NCPO', 'dateFormat' => 'Y-m']
-            );
-            $ncpo = RequestNCPO::create([
-                'ncpo_no' => $ncpoNo,
-                'date' => $validated['date'],
-                'justification' => $validated['justification'],
-                'po_id' => $validated['po_id'],
-                'created_by' => auth()->user()->id,
-                'approvals' => $validated['approvals'],
-            ]);
-            $ncpo->items()->createMany(
-                collect($validated['items'])->map(fn ($item) => [
-                    ...$item,
-                    'request_ncpo_id' => $ncpo->id,
-                ])->toArray()
-            );
-            return $ncpo->load('items');
-        });
-        if ($ncpo->getNextPendingApproval()) {
-            $ncpo->notify(new RequestNCPOForApprovalNotification($request->bearerToken(), $ncpo));
-        }
-        $ncpoResource = RequestNcpoResource::make($ncpo);
-        return $ncpoResource->additional([
-            'message' => 'Request NCPO created successfully.',
-            'success' => true,
-        ]);
-    }
+    // public function store(StoreNcpoRequest $request)
+    // {
+    //     $validated = $request->validated();
+    //     $ncpo = DB::transaction(function () use ($validated) {
+    //         $ncpoNo = RequestNCPO::generateReferenceNumber(
+    //             'ncpo_no',
+    //             fn ($prefix, $datePart, $number) => "{$prefix}-{$datePart}-{$number}",
+    //             ['prefix' => 'NCPO', 'dateFormat' => 'Y-m']
+    //         );
+    //         $ncpo = RequestNCPO::create([
+    //             'ncpo_no' => $ncpoNo,
+    //             'date' => $validated['date'],
+    //             'justification' => $validated['justification'],
+    //             'po_id' => $validated['po_id'],
+    //             'created_by' => auth()->user()->id,
+    //             'approvals' => $validated['approvals'],
+    //         ]);
+    //         $ncpo->items()->createMany(
+    //             collect($validated['items'])->map(fn ($item) => [
+    //                 ...$item,
+    //                 'request_ncpo_id' => $ncpo->id,
+    //             ])->toArray()
+    //         );
+    //         return $ncpo->load('items');
+    //     });
+    //     if ($ncpo->getNextPendingApproval()) {
+    //         $ncpo->notify(new RequestNCPOForApprovalNotification($request->bearerToken(), $ncpo));
+    //     }
+    //     $ncpoResource = RequestNcpoResource::make($ncpo);
+    //     return $ncpoResource->additional([
+    //         'message' => 'Request NCPO created successfully.',
+    //         'success' => true,
+    //     ]);
+    // }
 
     /**
      * Display the specified resource.
